Clean up the actor API
Also add a RenderableCullingMode so that we can create actors which don't get added to the partitioner - this is
necessary for debugging meshes etc.

Also needs adding to lights, directional lights should never be culled for example. Potentially also for particle systems
although I'm struggling to think of a case where that would matter...