WindowBase is no longer a ResourceManager
This is something I've been meaning to do for a while. Having windows and stages be asset managers
I thought gave a simple API, but really it was just confusing. The code was complex too, there had to be a
ResourceManager interface, which was implemented by Window, and Stage, and when it came to rewrite the UI
code and the UIStage needed to handle resources too it was time to rethink.

So, now window->shared_assets, stage->assets and uistage->assets are things, and they work as you'd expect. Assets
loaded into the stage live for the lifetime of the stage, assets loaded into the window live for the lifetime of the
window. You can access shared assets through the stage asset manager (just for ease).