Don't try to delete a Sprite's actor twice.
Sprites normally delete their actor's implicitly as the actor is a child node and all nodes are
destroyed recursively when a sprite is. This commit checks for existance of the actor before
trying to delete it in ~Sprite. Theoretically this isn't necessary, but it's probably better to be
safe than sorry.