Rename the project from KGLT (Kazade's OpenGL Toolkit) to Simulant
I've spent a long time trying to think of a name for this project. The initial name (KGLT) was only selected because originally
this wasn't a game engine, it was just a basic GL window toolkit. The name Simulant was chosen for a couple of reasons:

1. It means 'Something that simulates something else' and that's basically what a game engine does, it simulates worlds.
2. Simulants are one of the bad guys from Red Dwarf. I love Red Dwarf.

This name change marks the beginning of making this a proper usable game engine for other people, not just for myself. Now that I've settled on a name
I can build a website, write tutorials, hopefully gain contributors.

The engine is still pre-Alpha, but it's on its way!