Fix deadlocks when creating objects of the same type in multiple threads
Sometimes managed objects will directly or indirectly call code that needs to run on the main thread. In Simulant
that means queueing an idle task and waiting for the idle handler to run. The problem is that if the main thread
and another thread create an object (e.g. a stage) at around the same time. The other (non-main) thread may get the
StageManager lock, but then wait for the idle handler to run. The main thread will then block waiting for the
StageManager lock and not run the idle tasks - causing deadlocks.

This change instantiates the new object outside the lock, but doesn't acquire the manager lock until insertion into
the map, avoiding deadlocks.