Use Properties for VertexSpecification attributes
There was a massive performance issue when calling X_offset() on VertexSpecification, particularly
in the particle system which did it a lot, because these values were not pre-calculated - they couldn't be because
VertexSpecification was a struct with public members.

I've written a little VertexSpecificationProperty class which is like the generic Property<T, U> but calls recalc_stride_and_offsets()
when an attribute is set. This property is not used in performance critical code.