Add a default fps to KeyFrameAnimated
File formats like MS3D define a frame-rate for the whole model
rather than per-animation. So it makes sense to just allow a default
FPS, and allow specifying animations without explicitly specifying one.

The default value here is almost arbitrary, but 7FPS is common
in MD2 formats. Defaulting to zero would make it harder to work out
what was going on when the animations don't play at all. At least with
a non-zero default (whatever it is) it's clear to the user that the
animation is playing, but perhaps at the wrong speed.